Primary like Id's are managed by developer

+++ src/main/java/com/ecommerce/project/model/Category.java
@@ -1,19 +1,19 @@
 package com.ecommerce.project.model;
 
 public class Category {
-    private int categoryId;
+    private Long categoryId;
     private String categoryName;
 
-    public Category(int categoryId, String categoryName) {
+    public Category(Long categoryId, String categoryName) {
         this.categoryId = categoryId;
         this.categoryName = categoryName;
     }
 
-    public int getCategoryId() {
+    public Long getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(int categoryId) {
+    public void setCategoryId(Long categoryId) {
         this.categoryId = categoryId;
     }
